[metasrv] refactor: move raft types into metasrv/raft/types/
