+++ metasrv/src/meta_service/mod.rs
@@ -19,9 +19,6 @@ pub use log_entry::LogEntry;
 pub use meta_service_impl::MetaServiceImpl;
 pub use network::Network;
 pub use raft_txid::RaftTxId;
-pub use raft_types::LogIndex;
-pub use raft_types::NodeId;
-pub use raft_types::Term;
 pub use raftmeta::MetaNode;
 pub use raftmeta::MetaRaftStore;
 
@@ -38,15 +35,12 @@ pub mod log_entry;
 pub mod meta_service_impl;
 pub mod network;
 pub mod raft_txid;
-pub mod raft_types;
 pub mod raftmeta;
 
 #[cfg(test)]
 mod meta_service_impl_test;
 #[cfg(test)]
 mod meta_store_test;
-#[cfg(test)]
-mod raft_types_test;
 #[cfg(test)]
 pub mod raftmeta_test;
 #[cfg(test)]
