+++ metasrv/src/meta_service/raftmeta_test.rs
@@ -28,10 +28,10 @@ use crate::configs;
 use crate::meta_service::Cmd;
 use crate::meta_service::LogEntry;
 use crate::meta_service::MetaNode;
-use crate::meta_service::NodeId;
 use crate::meta_service::RaftTxId;
 use crate::meta_service::RetryableError;
 use crate::raft::state_machine::AppliedState;
+use crate::raft::types::NodeId;
 use crate::tests::assert_meta_connection;
 use crate::tests::service::new_test_context;
 use crate::tests::service::MetaSrvTestContext;
