+++ metasrv/src/raft/state_machine/sm.rs
@@ -45,7 +45,6 @@ use sled::IVec;
 use crate::configs;
 use crate::meta_service::Cmd;
 use crate::meta_service::LogEntry;
-use crate::meta_service::NodeId;
 use crate::raft::state_machine::placement::rand_n_from_m;
 use crate::raft::state_machine::AppliedState;
 use crate::raft::state_machine::Placement;
@@ -54,6 +53,7 @@ use crate::raft::state_machine::StateMachineMetaKey::Initialized;
 use crate::raft::state_machine::StateMachineMetaKey::LastApplied;
 use crate::raft::state_machine::StateMachineMetaKey::LastMembership;
 use crate::raft::state_machine::StateMachineMetaValue;
+use crate::raft::types::NodeId;
 use crate::sled_store::get_sled_db;
 use crate::sled_store::sled_key_space;
 use crate::sled_store::sled_key_space::StateMachineMeta;
