+++ metasrv/src/raft/types/mod.rs
@@ -11,8 +11,12 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+mod raft_types;
 
-pub mod log;
-pub mod state;
-pub mod state_machine;
-pub mod types;
+#[cfg(test)]
+mod raft_types_test;
+
+pub use raft_types::LogId;
+pub use raft_types::LogIndex;
+pub use raft_types::NodeId;
+pub use raft_types::Term;
