+++ metasrv/src/raft/types/raft_types_test.rs
@@ -14,7 +14,7 @@
 
 use std::ops::Bound;
 
-use crate::meta_service::NodeId;
+use crate::raft::types::NodeId;
 use crate::sled_store::sled_serde::SledOrderedSerde;
 use crate::sled_store::sled_serde::SledRangeSerde;
